pi/requirements.txt
# Raspberry Pi Master Server Dependencies
# Install with: pip install -r requirements.txt

# Web server + WebSocket
flask>=3.0
flask-socketio>=5.3
eventlet>=0.35

# I2C sensor support (Adafruit CircuitPython)
adafruit-blinka>=8.0
adafruit-circuitpython-apds9960>=3.0
adafruit-circuitpython-tca9548a>=0.7

# RFID reader
mfrc522>=0.0.7

# GPIO (pre-installed on Raspberry Pi OS, listed for completeness)
RPi.GPIO>=0.7
